Move numpy type to contrib as per pull comment.
